Use The CompetitionType Enum Instead Of Competition-Type Magic Numbers
Replaces the remaining bare 0/1 competition-type literals with CompetitionType
enum values. The read-only display DTOs (HomepageSeasonDto, SeasonDto,
MyLeagueDto, LeagueDashboardDto, ApiLeagueLookupResult) now carry the
CompetitionType enum directly, so their consumers compare against
CompetitionType.Tournament rather than `== 1`, and the dashboard state service
no longer casts.

Dapper maps the integer Competitions.Type column onto the enum automatically,
and enum values already round-trip over JSON between the API and client (as
RoundStatus does today), so this is behaviour-preserving.

The editable CompetitionDto.Type / BaseCompetitionRequest.Type form field is
left as an int: its comparisons already use enum values (cast to int), and
converting it would pull in an InputSelect binding change beyond this cleanup.

+++ src/ThePredictions.Application/Features/Admin/Seasons/Commands/CreateSeasonCommandHandler.cs
@@ -227,7 +227,7 @@ public class CreateSeasonCommandHandler(
             createdSeason.NumberOfRounds,
             competition.Id,
             competition.Name,
-            (int)competition.Type,
+            competition.Type,
             competition.ApiLeagueId,
             0, 0, 0, 0, 0, 0,
             createdSeason.PassStandardPrice,
